+++ app/index.js
@@ -1,7 +1,7 @@
 const config = require('./config');
 const slackWebApi = require('./infrastructure/slack-web-api');
 const createPostgreSQLDatabase = require('./infrastructure/postgresql/sequelize');
-const createExpressServer = require('./infrastructure/webserver/server');
+const createExpressServer = require('./infrastructure/express/server');
 const createPollsRepository = require('./interfaces/repositories/pollsRepositoryPostgreSQL');
 const createPollAnswersRepository = require('./interfaces/repositories/pollAnswersRepositoryPostgreSQL');
 const createPollsPresenter = require('./interfaces/presenters/pollsPresenterSlack');
